BREAKING CHANGE: chore: sync with internal lint (#75)
This synchronizes the lint rules to be closer to the lint rules used
internally at Google. Not all rules are available externally, so this
is not *exactly* the same.

Also updated the TypeScript version to 2.6.1.

A fair number of new lint rules are being applied now, which means that
the code will be alot more consistent with other Google code, but it
also means that users of gts may need to modify the source to pick up
this release. For most of the new lint rules, a fixer is also available
which should help with migration.

TypeScript 2.6.x enables contravariant types for function parameters,
which may also require some changes. See [1] for more info.

[1] https://blogs.msdn.microsoft.com/typescript/2017/10/31/announcing-typescript-2-6/