feat: upgrade to Typescript 2.8.x
Update the typescript dependency to 2.8.x both for the locally used dev
dependency and the version we emit in the generated package.json. The
peerDependency remains unchanged, which makes this a semver-minor.

Existing projects already have a dependency on typescript in their
package.json, so they should continue using the pre-2.8 version.

New versions projects will get 2.8.2.